+++ src/app/rundown-playlist/interfaces/rundown-playlist-service-interface.ts
@@ -1,5 +1,5 @@
 import {Observable} from 'rxjs';
-import {Identifier} from '../../shared/model/identifier';
+import {Identifier} from '../../shared/models/identifier';
 
 export interface RundownPlaylistServiceInterface {
   fetchRundownPlaylistIdentifiers(): Observable<Identifier[]>
