+++ src/app/rundown/interfaces/ad-lib-piece-service-interface.ts
@@ -1,5 +1,5 @@
 import {Observable} from 'rxjs';
-import {Identifier} from '../../shared/model/identifier';
+import {Identifier} from '../../shared/models/identifier';
 
 export interface AdLibPieceServiceInterface {
