+++ src/app/rundown/mocks/mock.ad-lib-piece.service.ts
@@ -1,6 +1,6 @@
 import {AdLibPieceServiceInterface} from '../interfaces/ad-lib-piece-service-interface';
 import {Observable, of} from 'rxjs';
-import {Identifier} from '../../shared/model/identifier';
+import {Identifier} from '../../shared/models/identifier';
 
 export class MockAdLibPieceService implements AdLibPieceServiceInterface {
