+++ src/app/shared/shared.module.ts
@@ -1,10 +1,10 @@
 import {NgModule} from '@angular/core'
 import {CommonModule} from '@angular/common'
-import {HttpErrorService} from './service/http-error.service'
 import {MAT_SNACK_BAR_DEFAULT_OPTIONS, MatSnackBarModule} from '@angular/material/snack-bar'
-import {RundownEventService} from './service/rundown-event.service';
 import {MockRundownEventService} from './mocks/mock.rundown-event.service';
 import {RundownEventServiceInterface} from './interfaces/rundown-event-service-interface';
+import {HttpErrorService} from './services/http-error.service';
+import {RundownEventService} from './services/rundown-event.service';
 
 const rundownEventService: RundownEventServiceInterface = new MockRundownEventService()
